+++ lib/fabricio/services/app_service.rb
@@ -12,7 +12,7 @@ module Fabricio
 
       # Initializes a new AppService object.
       #
-      # @param organization_id_provider [Fabricio::OrganizationIdProvider]
+      # @param organization_id_provider [Fabricio::Networking::OrganizationIdProvider]
       # @param network_client [Fabricio::Networking::NetworkClient]
       # @return [Fabricio::Service::AppService]
       def initialize(organization_id_provider, network_client)
